+++ Source/XV/Private/System/XVGameMode.cpp
@@ -1,9 +1,9 @@
-#include "XVGameMode.h"
-#include "XVGameState.h"
-#include "XVGameInstance.h"
-#include "ElevatorDoor.h"
+#include "System/XVGameMode.h"
+#include "System/XVGameState.h"
+#include "System/XVGameInstance.h"
+#include "World/ElevatorDoor.h"
 #include "Kismet/GameplayStatics.h"
-#include "SpawnVolume.h"
+#include "World/SpawnVolume.h"
 #include "AI/Character/Base/XVEnemyBase.h"
 #include "GameFramework/PlayerController.h"
